+++ src/mon.c
@@ -2500,9 +2500,9 @@ struct monst *mon;
 		struct obj *m_armr = which_armor(mon, W_ARM);
 
 		if (m_armr && Is_dragon_scales(m_armr))
-		    mndx = Dragon_scales_to_pm(m_armr) - mons;
+		    mndx = (int)(Dragon_scales_to_pm(m_armr) - mons);
 		else if (m_armr && Is_dragon_mail(m_armr))
-		    mndx = Dragon_mail_to_pm(m_armr) - mons;
+		    mndx = (int)(Dragon_mail_to_pm(m_armr) - mons);
 	      }
 		break;
 	}
